refactor function into helper function per PR comments
